flodoc v0.2.0
flodoc is a fork of flotate that adds support for JSDoc!

Most of the work here was wrestling with Esprima to find comments.
flotate's fancy-function-annotation feature is pretty inconsistent,
because where Esprima inserts precedingComments and trailingComments
is pretty inconsistent, so I had to write code to manually search
for preceding comments.